Choose same day when time is later than current
In cases such as "Monday 4pm", Chronic currently does not give the same
day when the time is later than what is given in `currently`. For
example, "Monday 4pm", when currently is a Monday and at 7am, will yield
_next_ Monday, but "Tuesday 4pm" will yield the following day.

This was a bit of a surprise and I've adjusted the code to first check
that if we are on the same day of week and the time is later to return
back the requested time for the current day rather than the next
matching day of week.