Fix parsing 'Sunday 12pm'
This checks the value from `Date.day_of_week` and maps it to the range
that Chronic is currently using, i.e.:

7  -> 0
x  -> x # 1-6

Unfortunately, this might mean there needs to be more done to retrofit
how Chronic is handling dates if it's out of line with what Elixir's
standard libraries are expecting. That or this should probably be
documented as an abnormality of the library, perhaps.